Trait salsa_2022::debug::DebugWithDb

pub trait DebugWithDb<Db: ?Sized> {
    // Required method
    fn fmt(
        &self,
        f: &mut Formatter<'_>,
        db: &Db,
        include_all_fields: bool
    ) -> Result;

    // Provided methods
    fn debug<'me, 'db>(&'me self, db: &'me Db) -> DebugWith<'me, Db>
       where Self: Sized + 'me { ... }
    fn debug_with<'me, 'db>(
        &'me self,
        db: &'me Db,
        include_all_fields: bool
    ) -> DebugWith<'me, Db>
       where Self: Sized + 'me { ... }
    fn debug_all<'me, 'db>(&'me self, db: &'me Db) -> DebugWith<'me, Db>
       where Self: Sized + 'me { ... }
    fn into_debug<'me, 'db>(self, db: &'me Db) -> DebugWith<'me, Db>
       where Self: Sized + 'me { ... }
    fn into_debug_all<'me, 'db>(self, db: &'me Db) -> DebugWith<'me, Db>
       where Self: Sized + 'me { ... }
}

Required Methods

fn fmt( &self, f: &mut Formatter<'_>, db: &Db, include_all_fields: bool ) -> Result

if include_all_fields is false only identity fields should be read, which means: - for #[salsa::input] no fields - for #[salsa::tracked] only fields with #[id] attribute - for #[salsa::interned] any field

Provided Methods

fn debug<'me, 'db>(&'me self, db: &'me Db) -> DebugWith<'me, Db>
where Self: Sized + 'me,

fn debug_with<'me, 'db>( &'me self, db: &'me Db, include_all_fields: bool ) -> DebugWith<'me, Db>
where Self: Sized + 'me,

fn debug_all<'me, 'db>(&'me self, db: &'me Db) -> DebugWith<'me, Db>
where Self: Sized + 'me,

Be careful when using this method inside a tracked function, because the default macro generated implementation will read all fields, maybe introducing undesired dependencies.

fn into_debug<'me, 'db>(self, db: &'me Db) -> DebugWith<'me, Db>
where Self: Sized + 'me,

fn into_debug_all<'me, 'db>(self, db: &'me Db) -> DebugWith<'me, Db>
where Self: Sized + 'me,

Be careful when using this method inside a tracked function, because the default macro generated implementation will read all fields, maybe introducing undesired dependencies.