Trait salsa_2022::ingredient::IngredientRequiresReset

pub trait IngredientRequiresReset {
    const RESET_ON_NEW_REVISION: bool;
}

Defines a const indicating if an ingredient needs to be reset each round. This const probably should be a member of Ingredient trait but then Ingredient would not be dyn-safe.

Required Associated Constants

const RESET_ON_NEW_REVISION: bool

If this is true, then reset_for_new_revision will be called every new revision.

Object Safety

This trait is not object safe.